Add xrandrwait implementation handling Crtc and OutputChange events
To dynamically handle changes in the configuration of connected screens,
an application that listens for XRandR events is needed.

This commit adds an initial implementation of the `xrandrwait` command,
which listens for XRandR events and writes them to standard output. The
command works in the same way as `inotifywait`. The caller can specify
the events that they are interested in using the `-e`/`--event` option,
or omit the option to have xrandrwait report all events.
If xrandrwait is started with `-m` or `--monitor`, it will listen for
events until it is interrupted by a signal. If the `-m`/`--monitor`
option is omitted, xrandrwait will terminate after the first event. If
a timeout is specified with the `-t` or `--timeout` options, xrandrwait
will terminate after the specified number of seconds if no event
occurred.

This version of xrandrwait can handle the following two event types:
 - XRROutputChangeNotifyEvent
 - XRRCrtcChangeNotifyEvent

Handling of `XRRScreenChangeNotifyEvent` messages is not implemented.